+++ insights/app/(main)/components/Sidebar.js
