+++ insights/app/(main)/components/SidebarCell.js
