+++ insights/app/(main)/components/SidebarCellContainer.js
