+++ insights/app/(main)/home/page.js
@@ -1,4 +1,4 @@
-import Header from "./components/Header";
+import Header from "../components/Header";
 
 export default function Dashboard() {
