+++ insights/app/(main)/layout.js
@@ -1,6 +1,5 @@
 import "../globals.css";
 import { Inter } from "next/font/google";
-import Header from "./components/Header";
 import Sidebar from "./components/Sidebar";
 
 const inter = Inter({ subsets: ["latin"], display: "swap" });
@@ -10,11 +9,15 @@ export const metadata = {
   description: "Made by Subodh Khanal",
 };
 
-export default function DashboardLayout({ children }) {
+export default function RootLayout({ children }) {
   return (
-    <main className="h-screen max-h-screen flex flex-row">
-      <Sidebar></Sidebar>
-      {children}
-    </main>
+    <html lang="en">
+      <body className={inter.classname}>
+        <main className="h-screen max-h-screen flex flex-row">
+          <Sidebar></Sidebar>
+          {children}
+        </main>
+      </body>
+    </html>
   );
 }
