+++ insights/app/(registration and login)/layout.js
@@ -1,0 +1,9 @@
+import '../globals.css'
+
+export default function LoginLayout({ children }) {
+  return (
+    <main>
+      {children}
+    </main>
+  )
+}
